+++ src/application/smalltalk/smalltalkSubject/dto/SmalltalkSubjectDto.ts
