+++ src/types/dropdown.d.ts
@@ -1,4 +1,4 @@
-import { SmalltalkSubjectDto } from "@/application/common/usecases/smalltalkSubject/dto/SmalltalkSubjectDto";
+import { SmalltalkSubjectDto } from "@/application/smalltalk/smalltalkSubject/dto/SmalltalkSubjectDto";
 
 export interface DropdownProps {
   options: SmalltalkSubjectDto[];
